+++ assets/components/romanescobackyard/js/caching/clear-cache-custom.js
@@ -1,0 +1,34 @@
+// Change partition variable to preferred cache path (relative to core/cache).
+// Subfolders can be used for more fine-grained cache controls, while still
+// being deleted when parent folder (custom) is cleared.
+var partition = 'custom';
+var topic = '/getcache/cache/partition/refresh/' + partition;
+
+this.console = MODx.load({
+    xtype: 'modx-console',
+    register: 'mgr',
+    topic: topic,
+    show_filename: 0
+});
+
+this.console.show(Ext.getBody());
+
+MODx.Ajax.request({
+    url: MODx.config.assets_url + 'components/getcache/connector.php',
+    params: {
+        action: 'cache/partition/refresh',
+        partitions: partition,
+        register: 'mgr',
+        topic: topic
+    },
+    listeners: {
+        'success': {
+            fn: function () {
+                this.console.fireEvent('complete');
+            }, scope: this
+        }
+    }
+});
+
+// noinspection JSAnnotator
+return false;